==> 1. ETL_MLOps.ipynb <==
# ETL

## Importación de librerias

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import ast
import gzip
import json
from textblob import TextBlob

## 1. Dataset 'steam_games.json. gz'

### Importación de datos del dataset

# Importación del Dataset
# Reemplaza 'mi_archivo.json.gz' con la ruta de tu archivo
with gzip.open('Dataset_inicial/steam_games.json.gz', mode='rt') as f:
    data = [json.loads(line) for line in f]

# Ponemos la variable como un DataFrame
games= pd.DataFrame(data)

### Muestra de los datos

games.head()

games.info(verbose=True)

# Cantidad de nulos por columnas
games.isnull().mean()*100

Se deduce que gran cantidad de los datos nulos del DataFrame afectaria considerablemente, sin embargo para esta ocasión se realizará la eliminación de los elementos nulos de manera sistematica

# Se eliminan las filas que no ofrecen información alguna
games.dropna(how= 'all', inplace= True)

# Registro num 88384 y 90890, en la mayoría no hay datos, podemos 
# eliminar estas filas
games.drop([88384,90890], inplace= True)

# Se puede eliminar la columna title por ser la misma a la 'app_name'
games.drop(columns= ['title'], inplace= True)

# Eliminamos las columnas que no ayudan a nuestro EDA
games.drop(columns= ['url', 'tags', 'reviews_url', 'early_access',
                     'specs'], inplace= True)


# Podemos realizar lo mismo para 'publisher' y 'developer'
condicion= games['publisher'].isnull() & games['developer'].notnull()
condicion2= games['publisher'].notnull() & games['developer'].isnull()

# Se reemplazan los valores nulos de las empresas que publicaron
# a desarrolladoras y viceversa
games.loc[condicion, 'publisher'] = games.loc[condicion, 'developer']
games.loc[condicion2, 'developer'] = games.loc[condicion2, 'publisher']

games.isnull().mean()*100

### Imputación de datos

Realizaremos una imputación de los valores nulos de los precios con los precios promedios de los generos que se asemejan

# Terminamos de eliminar el resto de elementos nulos
games.dropna(inplace=True)

# Se convierte la columna 'genres' para que sea hashable
games['genres']= games['genres'].apply(lambda x: tuple(x) if x is not None else x)

# La variable prom tendra las columnas genero y precios
prom= games.loc[:,['genres','price']]

# ponemos las variables que estan como lista a tuples para hashearlas
prom['genres']= prom['genres'].apply(lambda x: tuple(x))

# Elegimos todas los registros que tienen como variable float en la columna 'Price'
prom= prom.loc[prom['price'].apply(lambda x: not isinstance(x, str)), :]

# Se realiza la agrupación de los generos y se estima el promedio
prom['price'].astype(float)
promedio= prom.groupby('genres')['price'].mean()

# saca pone los indices en forma de columna
promedio= promedio.reset_index()
# promedio['genres'] = promedio.genres.apply(lambda x: list(x))
promedio.columns = ['genres','average price']

# Luego de realizar la imputación de los datos realizamos la conexion entre la tabla games y promedio y eliminamos los datos None
games= pd.merge(games, promedio, on= 'genres', how= 'left')
games['price']= games['price'].fillna(games['average price'])

# Reemplazamos los valores que son 'str' con los valores que sacamos de los promedios de la columna 'average price'
games['price'] = games.apply(lambda row: row['average price'] if isinstance(row['price'], str) else row['price'], axis=1)
games.dropna(inplace= True)

games['genres']= games['genres'].apply(lambda x: list(x))

### Exportación de datos

# Exportamos el DataFrame a archivo .parquet
games.to_parquet('Dataset_ETL/steam_games.parquet', engine='pyarrow')

## 2. Dataset 'user_reviews.json.gz'

### Importación los datos del dataset

# se crea una nueva lista en blanco para anexar los datos a esa lista
rows=[]

# con el metodo gzip, se lee el archivo .gz
with gzip.open('Dataset_inicial/user_reviews.json.gz', 'rb') as f:
    for line in f.readlines():
        rows.append(ast.literal_eval(line.decode(encoding= 'utf-8')))

# lo ponemos en un DataFrame llamado "reviews"
reviews= pd.DataFrame(rows)

### Exploración de los datos

reviews

Se ve que la columna 'reviews' tiene anidados datos en formato JSON, realizaremos su segmentación

df_review_explot

df_norm

# Descomponemos los datos estructurados de la lista de la columna 'reviews'
df_review_explot = reviews.explode('reviews')
# Normalizamos los datos en formato json y lo guardamos en un DataFrame
df_norm = pd.json_normalize(df_review_explot['reviews'])

# Iniciamos los indices desde cero
df_review_explot.reset_index(drop=True, inplace=True)
df_norm.reset_index(drop=True, inplace=True)

# Concatenamos los DataFrames
reviews = pd.concat([df_review_explot,df_norm], axis=1)

# Eliminamos la columna
reviews.drop(columns=['reviews'],inplace=True)

Cantidad de nulos en el DataFrame

reviews.isnull().mean()*100

reviews.info()

- Se realiza el borrado de todos los nulos por no ser tantos
- Eliminamos las columnas 'funny','last_edited','user_url','posted','helpful' por no contener datos relevantes para el analisis

# Se realiza el borrado de todos los nulos por no ser tantos
reviews.dropna(inplace=True)
reviews.reset_index(drop=True, inplace=True)

# Eliminamos las columnas 'funny','last_edited','user_url','posted','helpful' por no contener datos relevantes para el analisis
reviews.drop(columns=['funny','last_edited','user_url','posted','helpful'],inplace=True)

### Análisis de sentimientos

# Funcion para el analisis de sentimiento usando la libreria Textblob
def analyze_sentiment(review):
    if pd.isna(review) or review.strip() == '':
        return 1
    analysis = TextBlob(review)
    polarity = analysis.sentiment.polarity
    if polarity < -0.1:
        return 0
    elif polarity > 0.1:
        return 2
    else:
        return 1

- Si la polaridad es menor que -0.1, devuelve 0, lo que indica un sentimiento negativo.
- Si la polaridad es mayor que 0.1, devuelve 2, lo que indica un sentimiento positivo.
- Si la polaridad está entre -0.1 y 0.1 (inclusive), devuelve 1, lo que indica un sentimiento neutral.

# Aplicar la función de análisis de sentimiento
reviews['sentiment_analysis'] = reviews['review'].apply(analyze_sentiment)

# Reemplazar la columna 'review' con 'sentiment_analysis'
reviews = reviews.drop(columns=['review'])

# Mostrar el DataFrame resultante
print(reviews)

### Exportación de datos

reviews.to_parquet('Dataset_ETL/user_reviews.parquet',engine='pyarrow')

## 3. Dataset 'users_items.json.gz'

### Importación del dataset

rows=[]

with gzip.open('Dataset_inicial/users_items.json.gz', 'rb') as f:
    for line in f.readlines():
        rows.append(ast.literal_eval(line.decode(encoding= 'utf-8')))

items= pd.DataFrame(rows)



### Transformación, exploración y análisis del dataset

# Nomralizacion de la columna tipo JSON
df_items_explode = items.explode('items')
df_item_norm = pd.json_normalize(df_items_explode['items'])

# Inicializacion de indices
df_items_explode.reset_index(drop=True, inplace=True)
df_item_norm.reset_index(drop=True, inplace=True)

# Concatenizar los datasets
items = pd.concat([df_items_explode,df_item_norm], axis=1)

# Eliminamos la columna items
items.drop(columns=['items'], inplace=True)

# Normalizamos la columna item_name
items['item_name'] = items.item_name.str.title()

items.head()

items.info()

# Cantidad de nulos por columnas
items.isnull().sum()

condicion = items.playtime_forever.isnull()
items[condicion]

- Se ve que hay una cantidad pequeña de nulos en comparación del tamaño del dataset, se procederá a eliminar los nulos.
- Eliminamos las columnas que no seran relevantes para nuestro analisis en este caso serán 'user_url'

# eliminamos nulos, columna innecesaria y reseteamos los indices
items.dropna(inplace=True)
items.drop(columns=['user_url'],inplace=True)
items.reset_index(drop=True, inplace=True)

# Vemos si la columna 'user_id' podria ser la misma que 'steam_id' para eliminar alguna de ellas
(items.user_id == items.steam_id).value_counts()

# Pero no son iguales

items.shape

### Exportación del dataset

items.to_parquet('Dataset_ETL/users_items.parquet', engine='pyarrow')
