==> 2. Funciones_MLOps.ipynb <==
# Funciones: MLOps

## Importacion de las librerias y los datos

 # Importamos las librerias correspondientes
import pandas as pd
import numpy as np

# Leemos los dataset exportados con anterioridad
games= pd.read_parquet('Dataset_ETL/steam_games.parquet',engine='pyarrow')
reviews=pd.read_parquet('Dataset_ETL/user_reviews.parquet',engine='pyarrow')
items=pd.read_parquet('Dataset_ETL/users_items.parquet',engine='pyarrow')

games.head()

reviews.head()

items.head()

## Función 5: Developer Reviews Analysis
+ def **developer_reviews_analysis( *`desarrolladora` : str* )**:
    Según el desarrollador, se devuelve un diccionario con el nombre del desarrollador como llave y una lista con la cantidad total 
    de registros de reseñas de usuarios que se encuentren categorizados con un análisis de sentimiento como valor positivo o negativo. 

Ejemplo de retorno: {'Valve' : [Negative = 182, Positive = 278]}

# Se reañoza la acaptura de los datos
df_f5=pd.merge(games,reviews,how='inner',left_on='id',right_on='item_id')

# Importamos el dataset para el consumo de la funcion
df_f5.to_parquet('Dataset_funciones/df5_DeveloperRewiewsAnalysis.parquet',engine = 'pyarrow')

# Se filtra por desarrollador
dev = 'Valve'
df_dev = df_f5[df_f5.developer == dev]

# Se cuenta las reseñas positivas y nulas
df_dev = df_dev.sentiment_analysis.value_counts()

# Respuesta de la función
neg_ans = 'Negative = ' + str(df_dev[0])
pos_ans = 'Positive = ' + str(df_dev[2])
resp = {dev:[pos_ans, neg_ans]}

Funcion

def developer_reviews_analysis(desarrolladora:str):
    df_dev = df_f5[df_f5.developer == desarrolladora]
    df_dev = df_dev.sentiment_analysis.value_counts()
    neg_ans = 'Negative = ' + str(df_dev[0])
    pos_ans = 'Positive = ' + str(df_dev[2])
    resp = {desarrolladora:[pos_ans, neg_ans]}
    return resp

## Función 4: Best Developer Of Year
+ def **best_developer_year( *`año` : int* )**:
   Devuelve el top 3 de desarrolladores con juegos MÁS recomendados por usuarios para el año dado. (reviews.recommend = True y comentarios positivos)
   
Ejemplo de retorno: [{"Puesto 1" : X}, {"Puesto 2" : Y},{"Puesto 3" : Z}]

# Se reañoza la acaptura de los datos
df_f4=pd.merge(games,reviews,how='inner',left_on='id',right_on='item_id')

# Cambiamos el formato de las fechas por la correcta
df_f4['release_date'] = pd.to_datetime(df_f4.release_date, format='mixed')

# Creamos la columna 'year', para que se pueda delimitar el dataset
df_f4['year']=df_f4.release_date.dt.year

# Discriminamos el dataset por medio de todos los que tienen recomendaciones True
df_f4=df_f4[(df_f4.recommend==True) & (df_f4.sentiment_analysis>0)]

# Importamos el Dataset para la función 4
df_f4.to_parquet('Dataset_funciones/df4_BestDeveloperYear.parquet',engine = 'pyarrow')

df_f4.head()

df_f4_recommendations = df_f4[df_f4.year==1995]

df_f4_recommendations.head()

developer_recommendations = df_f4_recommendations['developer'].value_counts()

developer_recommendations = developer_recommendations.head(3).reset_index()

developer_recommendations

p1 = developer_recommendations.iloc[0,0]
p2 = developer_recommendations.iloc[1,0]
p3 = developer_recommendations.iloc[2,0]
resp = [{"Puesto 1" : p1}, {"Puesto 2" : p2},{"Puesto 3" : p3}]

resp

Funcion

def best_developer_year(año:int):
    df_f4_recommendations = df_f4[df_f4.year==año]
    developer_recommendations = df_f4_recommendations['developer'].value_counts()
    developer_recommendations = developer_recommendations.head(3).reset_index()
    p1 = developer_recommendations.iloc[0,0]
    p2 = developer_recommendations.iloc[1,0]
    p3 = developer_recommendations.iloc[2,0]
    resp = [{"Puesto 1" : p1}, {"Puesto 2" : p2},{"Puesto 3" : p3}]
    return resp

## Función 3: UseForGenre
+ def **UserForGenre( *`genero` : str* )**:
    Debe devolver el usuario que acumula más horas jugadas para el género dado y una lista de la acumulación de horas jugadas por año de lanzamiento.

Ejemplo de retorno: {"Usuario con más horas jugadas para Género X" : us213ndjss09sdf,
			     "Horas jugadas":[{Año: 2013, Horas: 203}, {Año: 2012, Horas: 100}, {Año: 2011, Horas: 23}]}

# Realizamos la captura de los datos a variables establcidas
df1_f3=games.copy()
df2_f3=items.copy()

# unimos los DataFrames captados
df3_f3=pd.merge(df2_f3,df1_f3,left_on='item_name',right_on='app_name',how='inner')

# Escogemos las columnas que solamente nos van a ayudar con el analisis
df3_f3=df3_f3.loc[:,['user_id','item_name','genres','playtime_forever','release_date']]
df3_f3

# Realizamos el desmenuzamiento de los generos
df3_f3=df3_f3.explode('genres')
# Reiniciamos los indices de la tabla
df3_f3.reset_index(drop=True,inplace=True)

# Damos el formato correcto a la columna de fechas
df3_f3['release_date']=pd.to_datetime(df3_f3['release_date'],format='%Y-%m-%d', errors='coerce')
# Se crea otra columna con los años de la columna con formato fecha
df3_f3['year']= df3_f3.release_date.dt.year

# Realizamos el agrupamiento y el ordenamiento
# Esto para conocer el usuario con mayor horas jugadas
df3_f3_agrupados=df3_f3.groupby(['genres','user_id'])['playtime_forever'].sum().reset_index()
df3_f3_ordenados=df3_f3_agrupados.sort_values(['genres','playtime_forever'],ascending=[True,False])
df3_f3_top=df3_f3_ordenados.drop_duplicates(subset='genres',keep='first')
df3_f3_top.reset_index(drop=True,inplace=True)
df3_f3_top.genres.replace('Animation &amp; Modeling','Animation & Modeling',inplace=True)
df3_f3_top.genres.replace('Design &amp; Illustration','Design & Illustration',inplace=True)

# Usuarios de cada genero con mayor cantidad de horas jugadas
df3_f3_top

Exportamos el primer dataset de la funcion

df3_f3_top.to_parquet('df3_UserForGenre_1.parquet')

# Realizamos el agrupamiento por genero, usuario y por año de las horas jugadas
df3_f3_agrupados_anio=df3_f3.groupby(['genres','user_id','year'])['playtime_forever'].sum().reset_index()

Exportamos el segundo dataset de la funcion

df3_f3_agrupados_anio.to_parquet('df3_UserForGenre_2.parquet')

# Se realiza una prueba para saber las horas jugadas por cada año
genero='Action'
usuario='shinomegami'
condicion=(df3_f3_agrupados_anio['genres'] == genero)&(df3_f3_agrupados_anio['user_id'] == usuario)
usuario_anios=df3_f3_agrupados_anio[condicion]
usuario_anios=usuario_anios.loc[:,['year','playtime_forever']].sort_values('year',ascending=False)
usuario_anios.columns=['Año','Horas']

d = usuario_anios.to_dict(orient='records')

df3_f3

Funcion

df3_1=pd.read_parquet('Dataset_funciones/df3_UserForGenre_1.parquet')
df3_2=pd.read_parquet('Dataset_funciones/df3_UserForGenre_2.parquet')

def UserForGenre(genero:str):
    condicion_1= (df3_1['genres'] == genero)
    seleccion=df3_1[condicion_1]
    usuario=seleccion.iloc[0,1]
    tiempo_total=seleccion.iloc[0,2]

    condicion_2=(df3_2['genres'] == genero)&(df3_2['user_id'] == usuario)
    tiempo_anio=df3_2[condicion_2]
    tiempo_anio=tiempo_anio.loc[:,['year','playtime_forever']].sort_values('year',ascending=False)
    tiempo_anio.columns=['Año','Horas']

    anios_dict = tiempo_anio.to_dict(orient='records')

    res={
        'Usuario con más horas jugadas para el Género ' + genero: usuario,
        'Hojas jugadas': anios_dict
    }

    return res

UserForGenre('RPG')

## Función 2: userdata
+ def **userdata( *`User_id` : str* )**:
    Debe devolver `cantidad` de dinero gastado por el usuario, el `porcentaje` de recomendación en base a reviews.recommend y `cantidad de items`.

Ejemplo de retorno: {"Usuario X" : us213ndjss09sdf, "Dinero gastado": 200 USD, "% de recomendación": 20%, "cantidad de items": 5}

df1=items.loc[:,['user_id','item_id','item_name','items_count']]
df2=reviews.loc[:,['user_id','recommend']]
df3=games.loc[:,['price','app_name','id']]

recomendacion=df2.groupby('user_id')['recommend'].value_counts(normalize=True)*100

recomendacion=recomendacion.reset_index()

recomendacion = recomendacion[recomendacion['recommend'] == True]

recomendacion['recommend'].value_counts()

recomendacion.tail()

df2.recommend.value_counts()

union_1=pd.merge(df1,df3,left_on="item_name",right_on="app_name",how="inner")

gastos = union_1.groupby('user_id')['price'].sum()
cant_items= union_1.groupby('user_id')['item_id'].count()

gastos
cant_items

gastos=gastos.reset_index()
cant_items=cant_items.reset_index()

df_userdata=pd.merge(gastos,cant_items,on='user_id',how='inner')
df_userdata=pd.merge(df_userdata,recomendacion,on='user_id',how='inner')

df_userdata.to_parquet('df2_userdata.parquet')

df_userdata.tail()

usuario=df_userdata[df_userdata['user_id'] == '-2SV-vuLB-Kg']

# usuario.price.iloc[0]

# gastos=int(usuario.price.iloc[0])
# gastos=str(gastos) + ' USD'

# recomendacion=str(int(usuario.proportion.iloc[0])) + '%'
# recomendacion

# cant_items=usuario.item_id.iloc[0]
# cant_items

num_usuario='Usuario '+str(usuario.index[0] + 1)
num_usuario

Funcion

def userdata(User_id:str):
    usuario=df_userdata[df_userdata['user_id'] == User_id]

    num_usuario='Usuario '+str(usuario.index[0] + 1)

    dinero_gastado=int(usuario.price.iloc[0])
    dinero_gastado=str(dinero_gastado) + ' USD'

    recomendacion=str(int(usuario.proportion.iloc[0])) + '%'

    cant_items=usuario.item_id.iloc[0]

    res={
        num_usuario: User_id,
        "Dinero gastado": dinero_gastado,
        "% de recomendación": recomendacion,
        "cantidad de items": cant_items
    }
    return res

userdata('-2SV-vuLB-Kg')

## Funcion 1: developer()
+ def **developer( *`desarrollador` : str* )**:
    `Cantidad` de items y `porcentaje` de contenido Free por año según empresa desarrolladora. 
Ejemplo de retorno:

| Año  | Cantidad de Items | Contenido Free  |
|------|-------------------|------------------|
| 2023 | 50                | 27%              |
| 2022 | 45                | 25%              |
| xxxx | xx                | xx%              |

Importación de datos

# Importamos el DataFrame a archivo .parquet
df_developer= games.copy()

# Se eliminan las variables que no ayudan a la formación de la funcion
df_developer.drop(columns=['publisher','app_name','price','average price'],inplace=True)

# Se realiza el cambio de formato de fecha sin contar con los errores
df_developer['release_date'] = pd.to_datetime(df_developer['release_date'],format='%Y-%m-%d', errors='coerce')

# Se crea otra columna con los años de la columna con formato fecha
df_developer['year']= df_developer.release_date.dt.year

# agregamos una columna que me haga conocer si hay FREE TO PLAY
df_developer['free']= df_developer['genres'].apply(lambda x: 'yes' if 'Free to Play' in x else 'no')

# Eliminamos genres y release_date por no ser necesarios al df de la funcion
df_developer.drop(columns=['genres','release_date'],inplace=True)

df_developer.head()

Exportación del dataset para la funcion 1

df_developer.to_parquet('df1_developer.parquet')

developer= pd.read_parquet('df1_developer.parquet')

developer

developer('Valve')

developer('Valve').to_dict(orient='records')

# Se toma otra variable con el nombre del desarrollador
games_anio_desa= games_anio[games_anio['developer'] == 'Valve']

# La cantidad de entregas que hay por año
games_anio_ordenado= games.groupby('year')['app_name'].count().sort_index(ascending=False)
games_anio_ordenado= games_anio_ordenado.reset_index()
games_anio_ordenado.columns= ['Año','Cantidad de Items']

contenido_free= games_anio_desa.groupby('year')['free'].value_counts(normalize=True).sort_index(ascending=False)*100
contenido_free= contenido_free.reset_index()
contenido_free= contenido_free[contenido_free['free'] == 'yes']
contenido_free.columns= ['Año','yes_no','Contenido Free']

resultado= pd.merge(games_anio_ordenado, contenido_free, on='Año', how='left')
resultado['Contenido Free']= resultado['Contenido Free'].fillna(0)
resultado['Contenido Free']= resultado['Contenido Free'].apply(lambda x: str(round(x,2)) + '%')
resultado['Año']= resultado['Año'].apply(lambda x: int(x))
resultado.drop(columns= ['yes_no'], inplace= True)

resultado

Función

def developer(desarrollador:str):
    # EXPLORATORY DATA ANALYSIS (EDA)
    # importamos los valores
    developer= pd.read_parquet('df1_developer.parquet')


    # obtenemos el dataframe segun el desarrollador
    developer= developer[developer['developer'] == desarrollador]

    # La cantidad de entregas que hay por año
    games_anio_ordenado= developer.groupby('year')['id'].count().sort_index(ascending=False)
    games_anio_ordenado= games_anio_ordenado.reset_index()  # se convierte los indices en columnas 
    games_anio_ordenado.columns= ['Año','Cantidad de Items']  # cambiamos el nombre de las columnas



    # realizamos la agrupacion y contamos los elemnentos por año y por ser 'Free to Play' 
    contenido_free= developer.groupby('year')['free'].value_counts(normalize=True).sort_index(ascending=False)*100
    contenido_free= contenido_free.reset_index()  # Se colocan los indices como columnas
    contenido_free= contenido_free[contenido_free['free'] == 'yes']  # Se realiza el filtro con solo los 'yes'
    contenido_free.columns= ['Año','yes_no','Contenido Free']  # Se cambia el nombre de las columnas
    

    # realizamos el juntar la tabla de las cantidad de items y porcentaje de contenido Free
    # tambien se realizan otras modificaciones al llenar con ceros los 'NaN', eliminacion de la columna 'yes_no'
    resultado= pd.merge(games_anio_ordenado, contenido_free, on='Año', how='left')
    resultado['Contenido Free']= resultado['Contenido Free'].fillna(0)
    resultado['Contenido Free']= resultado['Contenido Free'].apply(lambda x: str(round(x,2)) + '%')
    resultado['Año']= resultado['Año'].apply(lambda x: int(x))
    resultado.drop(columns= ['yes_no'], inplace= True)
    
    return resultado
