==> 4. Recomendation_System.ipynb <==
# Sistema de recomendacion

## Importación de librerias

 # Importamos las librerias correspondientes
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

## Importación de datasets

# Leemos los dataset exportados con anterioridad
games= pd.read_parquet('Dataset_ETL/steam_games.parquet',engine='pyarrow')
reviews=pd.read_parquet('Dataset_ETL/user_reviews.parquet',engine='pyarrow')
items=pd.read_parquet('Dataset_ETL/users_items.parquet',engine='pyarrow')

games.head()

reviews.head()

recommends = pd.merge(reviews,games,left_on='item_id',right_on='id')

recommends=recommends.iloc[:,[0,1,3,6]]

recommends = recommends[recommends.sentiment_analysis == 1]

recommends

recommends.info()

# Convertir la escala 2, 1, 0 a 1, 0, -1
recommends.sentiment_analysis = recommends.sentiment_analysis.map({2:1,1:0,0:-1})

# Concatenar los valores duplicados en una cadena
df_pivot = recommends.pivot_table(index='user_id', columns='app_name', values='sentiment_analysis', aggfunc='mean')

df_pivot.fillna(0,inplace=True)

df_pivot

print(df_pivot.iloc[9,:].value_counts())

print(df_pivot.iloc[0,:].dtype)

df_pivot.iloc[:,0].astype('int')

df_pivot.info()

He podido realizar la tabla con los datos de los usuarios y con la parte de las calificaciones
Toca realizar en el dataset recommends, la similitud del coseno, esto hara que los usuarios tengan recomendaciones de juegos similares con una lista de 5 juegos similares que puedan gustarle.

## Similitud del coseno

from sklearn.metrics.pairwise import cosine_similarity
import pandas as pd

# Asumamos que 'df_pivot' es tu tabla de usuarios y juegos con valores de sentiment_analysis
# Asegúrate de que 'user_id' sea el índice del DataFrame
# df_pivot = df_pivot.set_index('user_id')

# Calculamos la similitud del coseno entre todos los usuarios
user_similarity = cosine_similarity(df_pivot)

# Convertimos la matriz de similitud en un DataFrame para que sea más fácil manejarlo
user_similarity_df = pd.DataFrame(user_similarity, index=df_pivot.index, columns=df_pivot.index)

# Ver la matriz de similitud
print(user_similarity_df)

# Elegimos el usuario para el que queremos hacer la recomendación
target_user = '--000--'  # Ejemplo con el usuario 1

# Ordenamos las similitudes del usuario en orden descendente
similar_users = user_similarity_df.loc[target_user].sort_values(ascending=False)

# Excluimos al propio usuario (ya que la similitud consigo mismo es 1)
similar_users = similar_users.drop(target_user)

# Vemos los usuarios más similares
print(similar_users.head())

# Obtenemos los juegos que el usuario objetivo ya ha calificado
user_games = df_pivot.loc[target_user]

# Filtramos los juegos que el usuario ya ha jugado o que no le gustaron (valores 0 o negativos)
games_user_played = user_games[user_games > 0].index

# Ahora buscamos los juegos que han disfrutado los usuarios más similares
# Seleccionamos los 5 usuarios más similares
top_similar_users = similar_users.head(5).index

# Filtramos los juegos que estos usuarios disfrutaron (con puntuaciones positivas)
recommended_games = df_pivot.loc[top_similar_users].mean().sort_values(ascending=False)

# Excluimos los juegos que ya jugó el usuario
recommended_games = recommended_games.drop(games_user_played)

# Recomendamos los 5 juegos mejor valorados entre usuarios similares
top_5_recommendations = recommended_games.head(5)

print("Juegos recomendados para el usuario:", list(top_5_recommendations.index))

Exportamos los datasets "user_similarity_df" que es el dataset que esta calculada la similitud del coseno en los usuarios.

Y exportamos el dataset "df_pivot" donde esta la tabla pivote que hemos obtenido para el analisis de las recomendacion de los usuarios.

df_pivot.to_parquet('Dataset_funciones/df_recomendacion_PivotTable.parquet',engine = 'pyarrow')
user_similarity_df.to_parquet('Dataset_funciones/df_recomendacion_UserSimilarity.parquet',engine = 'pyarrow')

df_UserSimilarity=pd.read_parquet('./Dataset_funciones/df_recomendacion_UserSimilarity.parquet',engine='pyarrow')

df_UserSimilarity
